refactor(kotlin): split AnnotationAttributeAnnotator into two classes
Extract function-parameter type annotation into a dedicated class,
separating the two unrelated responsibilities.

- New FunctionParameterAnnotator: setFunctionParameterTypes +
  annotateParameterNodes (moved verbatim from AnnotationAttributeAnnotator)
- AnnotationAttributeAnnotator: annotation FQN logic only; remove
  KtFunctionDeclaration, KtFunctionValueParameters, ParameterAst imports;
  update javadoc
- KotlinTypeAnnotationVisitor: call FunctionParameterAnnotator instead of
  AnnotationAttributeAnnotator for parameter types; update javadoc
